requirements.txt
asgiref==3.2.10
Django==3.0.8
django-braces==1.14.0
django-crispy-forms==1.9.2
django-model-utils==4.0.0
factory-boy==2.12.0
Faker==4.1.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
text-unidecode==1.3
